rename: 미션 제공 테스트코드 네이밍 수정
