+++ src/test/java/roomescape/DatabaseStepTest.java
@@ -1,9 +1,10 @@
-package roomescape.domain;
+package roomescape;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
 import java.sql.Connection;
 import java.sql.SQLException;
+import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.context.SpringBootTest;
@@ -18,7 +19,8 @@ public class DatabaseStepTest {
     private JdbcTemplate jdbcTemplate;
 
     @Test
-    void 사단계() {
+    @DisplayName("DB 커넥션을 테스트 한다")
+    void db_connection() {
         try (Connection connection = jdbcTemplate.getDataSource().getConnection()) {
             assertThat(connection).isNotNull();
             assertThat(connection.getCatalog()).isEqualTo("DATABASE-TEST");
